compiler: Fix bug in embedded method handling.
If the same type was embedded in a struct at different levels,
only the first embedded instance was considered when building
the set of all methods.  A second instance was ignored, even
if it was at a higher level and should therefore override the
first instance.

Test case is bug485.go in master repository.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@210096 138bc75d-0d04-0410-961f-82ee72b054a4